backend/requirements.txt
# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20

# LLM Integration
huggingface-hub>=0.20.0
openai>=1.0.0

# FastAPI and Web Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database (MongoDB)
motor>=3.3.0
pymongo>=4.6.0

# Data Validation and Serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Requests and APIs
requests>=2.31.0
httpx>=0.25.0
aiohttp>=3.9.0

# Environment and Configuration
python-dotenv>=1.0.0

# Document Generation
python-docx>=1.1.0
docx2txt>=0.8

# Utilities
pyyaml>=6.0
typing-extensions>=4.8.0

# Date and Time
python-dateutil>=2.8.2

# Background Tasks and Async
celery>=5.3.0
redis>=5.0.0

# Testing (Optional but recommended)
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Code Quality (Optional)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
